// src/menus/application/dto/create-menu.dto.ts
export interface CreateMenuItemDto {
  label: string;
  icon?: string;
  order?: number;
  entityDefinitionId?: string;
  path?: string;
}

export interface CreateMenuGroupDto {
  name: string;
  order?: number;
  items?: CreateMenuItemDto[];
}

export interface CreateMenuDto {
  tenantId: string;
  name: string;
  icon?: string;
  order: number;
  isActivate: boolean;
  groups: CreateMenuGroupDto[];
}